+++ bin/vigild/src/collector/mod.rs
@@ -5,6 +5,7 @@ mod manager;
 mod run;
 #[cfg(test)]
 mod tests;
+mod waiting;
 
 pub(crate) use host::SYSTEMCTL_PLACES;
 pub use run::{Options, disable, enable};