+++ bin/vigild/src/collector/tests/mod.rs
@@ -1,2 +1,3 @@
 mod apart;
 mod edit;
+mod waiting;